backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
nltk==3.9.1
numpy==1.26.4
